+++ mse_cli/command/logout.py
@@ -18,6 +18,12 @@ def add_subparser(subparsers):
 
 def run(_args) -> None:
     """Run the subcommand."""
+
+    logout()
+    LOG.success("You are now logged out.")  # type: ignore
+
+
+def logout() -> None:
     r = requests.post(url=f"{MSE_AUTH0_DOMAIN_NAME}/logout?federated", timeout=60)
 
     if not r.ok:
@@ -26,5 +32,3 @@ def run(_args) -> None:
     login_file = UserConf.path()
     if login_file.exists():
         os.remove(login_file)
-
-    LOG.success("You are now logged out.")  # type: ignore