+++ mse_cli/__init__.py
@@ -4,7 +4,7 @@ import os
 import sys
 from pathlib import Path
 
-__version__ = "0.11.0"
+__version__ = "0.12.0"
 
 # This directory contains the login information of the user
 # and the context of all its deployments