+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * `mse list` does not required the project name. It's optional.
 * Use `get_server_certificate` from `intel-sgx-ra`
 * Support refresh token expiration
+* Take the default configuration values form the backend 
 
 ## \[0.10.2\] - 2023-03-01